+++ node/deploy_config.ts
@@ -2,10 +2,9 @@ import { promises as fs } from 'fs'
 import { dirname, resolve } from 'path'
 
 import type { Declaration } from './declaration.js'
-import { ImportMapFile, readFile as readImportMap } from './import_map.js'
 import type { Layer } from './layer.js'
 import type { Logger } from './logger.js'
-import { isNodeError } from './utils/error.js'
+import { isFileNotFoundError } from './utils/error.js'
 
 /* eslint-disable camelcase */
 interface DeployConfigFile {
@@ -18,7 +17,7 @@ interface DeployConfigFile {
 
 export interface DeployConfig {
   declarations: Declaration[]
-  importMap?: ImportMapFile
+  importMap?: string
   layers: Layer[]
 }
 
@@ -36,7 +35,7 @@ export const load = async (path: string | undefined, logger: Logger): Promise<De
 
     return parse(config, path)
   } catch (error) {
-    if (isNodeError(error) && error.code !== 'ENOENT') {
+    if (!isFileNotFoundError(error)) {
       logger.system('Error while parsing internal edge functions manifest:', error)
     }
   }
@@ -47,24 +46,20 @@ export const load = async (path: string | undefined, logger: Logger): Promise<De
   }
 }
 
-const parse = async (data: DeployConfigFile, path: string): Promise<DeployConfig> => {
+const parse = (data: DeployConfigFile, path: string) => {
   if (data.version !== 1) {
     throw new Error(`Unsupported file version: ${data.version}`)
   }
 
-  const config = {
+  const config: DeployConfig = {
     declarations: data.functions ?? [],
     layers: data.layers ?? [],
   }
 
   if (data.import_map) {
     const importMapPath = resolve(dirname(path), data.import_map)
-    const importMap = await readImportMap(importMapPath)
 
-    return {
-      ...config,
-      importMap,
-    }
+    config.importMap = importMapPath
   }
 
   return config