+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [7.0.1](https://github.com/netlify/edge-bundler/compare/v7.0.0...v7.0.1) (2022-12-19)
+
+
+### Bug Fixes
+
+* read other import map paths when one doesn't exist ([#268](https://github.com/netlify/edge-bundler/issues/268)) ([8410c1c](https://github.com/netlify/edge-bundler/commit/8410c1caf9c57edf9ebb6874dd06750626d94a8d))
+
+## [7.0.0](https://github.com/netlify/edge-bundler/compare/v6.1.0...v7.0.0) (2022-12-19)
+
+
+### ⚠ BREAKING CHANGES
+
+* Both `bundle` and `serve` now expect an `importMapPaths` array containing a list of paths to any user-defined import map files
+
+### Features
+
+* add `importMapPaths` to `bundle` and `serve` ([#265](https://github.com/netlify/edge-bundler/issues/265)) ([dcbd7f7](https://github.com/netlify/edge-bundler/commit/dcbd7f7681f8971bd88e251669b7f09602ee0b9c))
+* log failures when loading import map files ([#267](https://github.com/netlify/edge-bundler/issues/267)) ([138690b](https://github.com/netlify/edge-bundler/commit/138690b5e490b5d577f07c529bcbfb1379be4811))
+
 ## [6.1.0](https://github.com/netlify/edge-bundler/compare/v6.0.0...v6.1.0) (2022-12-14)